backend/requirement.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart>=0.0.5
assemblyai
google-generativeai